+++ pkg/controller/oneagent/oneagent_controller_test.go
@@ -165,6 +165,7 @@ func TestReconcile_PhaseSetCorrectly(t *testing.T) {
 		// arrange
 		oa := base.DeepCopy()
 		oa.Status.Version = version
+		oa.Status.Tokens = utils.GetTokensName(oa)
 
 		// act
 		updateCR, err := reconciler.reconcileRollout(logger, oa, dtcMock)
@@ -186,6 +187,84 @@ func TestReconcile_PhaseSetCorrectly(t *testing.T) {
 		// assert
 		assert.Equal(t, nil, err)
 		assert.Equal(t, dynatracev1alpha1.OneAgentPhaseType(""), oa.Status.Phase)
+	})
+}
+
+func TestReconcile_TokensSetCorrectly(t *testing.T) {
+	namespace := "dynatrace"
+	oaName := "oneagent"
+	base := dynatracev1alpha1.OneAgent{
+		ObjectMeta: metav1.ObjectMeta{Name: oaName, Namespace: namespace},
+		Spec: dynatracev1alpha1.OneAgentSpec{
+			BaseOneAgentSpec: dynatracev1alpha1.BaseOneAgentSpec{
+				APIURL: "https://ENVIRONMENTID.live.dynatrace.com/api",
+				Tokens: oaName,
+			},
+		},
+	}
+	logger := logf.ZapLoggerTo(os.Stdout, true)
+	c := fake.NewFakeClientWithScheme(scheme.Scheme, NewSecret(oaName, namespace, map[string]string{utils.DynatracePaasToken: "42", utils.DynatraceApiToken: "84"}))
+	dtcMock := &dtclient.MockDynatraceClient{}
+	version := "1.187"
+	dtcMock.On("GetLatestAgentVersion", dtclient.OsUnix, dtclient.InstallerTypeDefault).Return(version, nil)
+
+	reconciler := &ReconcileOneAgent{
+		client:    c,
+		apiReader: c,
+		scheme:    scheme.Scheme,
+		logger:    logf.ZapLoggerTo(os.Stdout, true),
+		dtcReconciler: &utils.DynatraceClientReconciler{
+			Client:              c,
+			DynatraceClientFunc: utils.StaticDynatraceClient(dtcMock),
+			UpdatePaaSToken:     true,
+			UpdateAPIToken:      true,
+		},
+	}
+
+	t.Run("reconcileRollout Tokens status set, if empty", func(t *testing.T) {
+		// arrange
+		oa := base.DeepCopy()
+		oa.Spec.Tokens = ""
+		oa.Status.Tokens = ""
+
+		// act
+		updateCR, err := reconciler.reconcileRollout(logger, oa, dtcMock)
+
+		// assert
+		assert.True(t, updateCR)
+		assert.Equal(t, utils.GetTokensName(oa), oa.Status.Tokens)
+		assert.Equal(t, nil, err)
+	})
+	t.Run("reconcileRollout Tokens status set, if status has wrong name", func(t *testing.T) {
+		// arrange
+		oa := base.DeepCopy()
+		oa.Spec.Tokens = ""
+		oa.Status.Tokens = "not the actual name"
+
+		// act
+		updateCR, err := reconciler.reconcileRollout(logger, oa, dtcMock)
+
+		// assert
+		assert.True(t, updateCR)
+		assert.Equal(t, utils.GetTokensName(oa), oa.Status.Tokens)
+		assert.Equal(t, nil, err)
+	})
+
+	t.Run("reconcileRollout Tokens status set, not equal to defined name", func(t *testing.T) {
+		// arrange
+		customTokenName := "custom-token-name"
+		oa := base.DeepCopy()
+		oa.Status.Tokens = utils.GetTokensName(oa)
+		oa.Spec.Tokens = customTokenName
+
+		// act
+		updateCR, err := reconciler.reconcileRollout(logger, oa, dtcMock)
+
+		// assert
+		assert.True(t, updateCR)
+		assert.Equal(t, utils.GetTokensName(oa), oa.Status.Tokens)
+		assert.Equal(t, customTokenName, oa.Status.Tokens)
+		assert.Equal(t, nil, err)
 	})
 }